move some utils up out of baseline to 8mi
