+++ layers/eight_mile/pytorch/layers.py
@@ -3498,3 +3498,59 @@ class TransformerDiscriminator(nn.Module):
                 real_loss = self.loss(input[target != 0], target[target != 0])
                 return real_loss + fake_loss
         return Loss()
+
+
+class SequenceCriterion(nn.Module):
+
+    def __init__(self, LossFn=nn.NLLLoss, avg='token'):
+        super(SequenceCriterion, self).__init__()
+        if avg == 'token':
+            # self.crit = LossFn(ignore_index=Offsets.PAD, reduction='elementwise-mean')
+            self.crit = LossFn(ignore_index=Offsets.PAD, size_average=True)
+            self._norm = self._no_norm
+        else:
+            self.crit = LossFn(ignore_index=Offsets.PAD, size_average=False)
+            self._norm = self._batch_norm
+
+    def _batch_norm(self, loss, inputs):
+        return loss / inputs.size()[0]
+
+    def _no_norm(self, loss, inputs):
+        return loss
+
+    def forward(self, inputs, targets):
+        """Evaluate some loss over a sequence.
+
+        :param inputs: torch.FloatTensor, [B, .., C] The scores from the model. Batch First
+        :param targets: torch.LongTensor, The labels.
+
+        :returns: torch.FloatTensor, The loss.
+        """
+        total_sz = targets.nelement()
+        loss = self.crit(inputs.view(total_sz, -1), targets.view(total_sz))
+        return self._norm(loss, inputs)
+
+
+def pytorch_conv1d(in_channels, out_channels, fsz, unif=0, padding=0, initializer=None):
+    c = nn.Conv1d(in_channels, out_channels, fsz, padding=padding)
+    if unif > 0:
+        c.weight.data.uniform_(-unif, unif)
+    elif initializer == "ortho":
+        nn.init.orthogonal(c.weight)
+    elif initializer == "he" or initializer == "kaiming":
+        nn.init.kaiming_uniform(c.weight)
+    else:
+        nn.init.xavier_uniform_(c.weight)
+    return c
+
+
+def tie_weight(to_layer, from_layer):
+    """Assigns a weight object to the layer weights.
+
+    This method exists to duplicate baseline functionality across packages.
+
+    :param to_layer: the pytorch layer to assign weights to
+    :param from_layer: pytorch layer to retrieve weights from
+    """
+    to_layer.weight = from_layer.weight
+
