+++ layers/eight_mile/tf/embeddings.py
@@ -467,44 +467,3 @@ class LearnedPositionalCharLSTMEmbeddings(LearnedPositionalMixin, CharLSTMEmbedd
         T = tf.shape(x)[1]
         pos = self.positional(T)
         return self.dropout(x + pos, training=TRAIN_FLAG())
-
-
-# All the train functions don't have the large lut init codes anymore and it needs a placeholder
-# So lets skip it for now
-# https://stackoverflow.com/questions/43288147/how-do-i-use-a-very-large-2m-word-embedding-in-tensorflow
-# class LargeLookupTableEmbeddings(LookupTableEmbeddings):
-#     """Create a really large embedding matrix in tensorflow.
-
-#     Tensorflow has a limit on the size that a op can be (2GB). When we have very
-#     large embedding lookuptables (for example when we don't prune the vocab) we
-#     hit this limit and can't have the embeddings in the graph. This is due to a
-#     limit in the size that a thing can be in a protocol buffer (how tensorflow
-#     serializes the graph).
-
-#     Here we get around it with a placeholder. The place holder will be in the
-#     graph and it will know it needs to have a size of [vsz, dsz] but it doesn't
-#     have the actual values so it can be serialized into a protocol buffer since
-#     it is small.
-
-#     We then have a variable that is initialized with the value of the
-#     placeholder. This is filled in with value during the `sess.run` of
-#     `tf.compat.v1.global_variables_initialzier` with a feed_dict. Values are then saved
-#     into the checkpoint and can be reloaded from there.
-
-#     ```
-#     sess.run(tf.compat.v1.global_variables_initializer(), {e.W_place: e._weights})
-#     ```
-
-#     The future of this with tf 2 is unclear because it uses a placeholder to
-#     get around the size limit and placeholders are not cool in tf2
-#     """
-#     def __init__(self, trainable=True, name=None, dtype=tf.float32, **kwargs):
-#         super().__init__(trainable, name, dtype, **kwargs)
-#         self.W_place = tf.placeholder(tf.float32, shape=(self.vsz, self.dsz))
-
-#     def build(self, input_shape):
-#         self.W = tf.get_variable(f"{self.scope}/weight", initializer=self.W_place)
-
-#     def get_feed_dict(self):
-#         """Feed dict mapping the numpy weights to the placeholder."""
-#         return {self.W_place: self._weights}
