+++ layers/eight_mile/tf/layers.py
@@ -1,13 +1,14 @@
 import logging
 import tensorflow as tf
 import numpy as np
-from eight_mile.utils import listify, Offsets, wraps, get_version, is_sequence
+from eight_mile.utils import listify, Offsets, wraps, get_version, is_sequence, transition_mask as transition_mask_np
 from typing import Optional, Union, List, Dict, Any, Tuple
 import contextlib
 import math
 
 BASELINE_TF_TRAIN_FLAG = None
 LOGGER = logging.getLogger('mead.layers')
+
 
 @contextlib.contextmanager
 def autograph_options(options):
@@ -17,6 +18,7 @@ def autograph_options(options):
         yield
     finally:
         tf.config.optimizer.set_experimental_options(old_opts)
+
 
 def set_tf_eager_mode(prefer_eager: bool = False):
     tf_version = get_version(tf)
@@ -206,7 +208,7 @@ class WithoutLength(tf.keras.layers.Layer):
         output = self.layer(inputs[0])
         return output
 
-# Mapped
+
 class ConvEncoder(tf.keras.layers.Layer):
     def __init__(self, insz: Optional[int], outsz: int, filtsz: int, pdrop: float = 0.0, activation: str = "relu", name=None):
         super().__init__(name=name)
@@ -3129,6 +3131,88 @@ def reload_checkpoint(sess: tf.compat.v1.Session, checkpoint: str, blocks_to_ski
     LOGGER.info("Restoring %s", g)
     saver = tf.compat.v1.train.Saver(g)
     saver.restore(sess, latest)
+
+
+def reload_embeddings(embeddings_dict, basename):
+    embeddings = {}
+    for key, cls in embeddings_dict.items():
+        embed_args = read_json('{}-{}-md.json'.format(basename, key))
+        module = embed_args.pop('module')
+        name = embed_args.pop('name', None)
+        assert name is None or name == key
+        mod = import_user_module(module)
+        embed_args['name'] = key
+        Constructor = getattr(mod, cls)
+        embeddings[key] = Constructor(**embed_args)
+    return embeddings
+
+
+
+def transition_mask(vocab, span_type, s_idx, e_idx, pad_idx=None):
+    """Create a CRF Mask.
+    Returns a mask with invalid moves as 0 and valid moves as 1.
+    """
+    mask = transition_mask_np(vocab, span_type, s_idx, e_idx, pad_idx).T
+    inv_mask = (mask == 0).astype(np.float32)
+    return mask, inv_mask
+
+
+def tie_weight(weight, tie_shape):
+    """Higher order function to share weights between two layers.
+
+    Tensorflow will take a custom_getter inside of a variable scope.
+    This method creates a getter that looks for a match in shapes. If they match,
+    The weights are transposed and shared.
+
+    """
+    def tie_getter(getter, name, *args, **kwargs):
+        if kwargs['shape'] == tie_shape:
+            return tf.transpose(weight)
+        return getter("{}".format(name), *args, **kwargs)
+    return tie_getter
+
+
+def rnn_cell_w_dropout(hsz, pdrop, rnntype, st=None, variational=False, training=False):
+
+    """Produce a single RNN cell with dropout
+    :param hsz: (``int``) The number of hidden units per LSTM
+    :param rnntype: (``str``): `lstm` or `gru`
+    :param pdrop: (``int``) The probability of dropping a unit value during dropout
+    :param st: (``bool``) state is tuple? defaults to `None`
+    :param variational: (``bool``) Variational recurrence is on
+    :param training: (``bool``) Are we training?  Defaults to ``False``
+    :return: a cell
+    """
+    output_keep_prob = tf.contrib.framework.smart_cond(training, lambda: 1.0 - pdrop, lambda: 1.0)
+    state_keep_prob = tf.contrib.framework.smart_cond(training, lambda: 1.0 - pdrop if variational else 1.0, lambda: 1.0)
+    cell = rnn_cell(hsz, rnntype, st)
+    output = tf.contrib.rnn.DropoutWrapper(cell,
+                                           output_keep_prob=output_keep_prob,
+                                           state_keep_prob=state_keep_prob,
+                                           variational_recurrent=variational,
+                                           dtype=tf.float32)
+    return output
+
+
+def multi_rnn_cell_w_dropout(hsz, pdrop, rnntype, num_layers, variational=False, training=False):
+    """Produce a stack of RNNs with dropout performed on all but the last layer.
+
+    :param hsz: (``int``) The number of hidden units per RNN
+    :param pdrop: (``int``) The probability of dropping a unit value during dropout
+    :param rnntype: (``str``) The type of RNN to use - `lstm` or `gru`
+    :param num_layers: (``int``) The number of layers of RNNs to stack
+    :param training: (``bool``) Are we training? Defaults to ``False``
+    :return: a stacked cell
+    """
+    if variational:
+        return tf.contrib.rnn.MultiRNNCell(
+            [rnn_cell_w_dropout(hsz, pdrop, rnntype, variational=variational, training=training) for _ in range(num_layers)],
+            state_is_tuple=True
+        )
+    return tf.contrib.rnn.MultiRNNCell(
+        [rnn_cell_w_dropout(hsz, pdrop, rnntype, training=training) if i < num_layers - 1 else rnn_cell_w_dropout(hsz, 1.0, rnntype) for i in range(num_layers)],
+        state_is_tuple=True
+    )
 
 
 def tf_device_wrapper(func):
